+++ src/frontend/src/types/index.ts
@@ -1,3 +1,4 @@
 export * from "./auth";
 export * from "./s3";
 export * from "./nft";
+export * from "./dashboard";